requirements.txt
# ── Core LLM Framework ─────────────────────────────────────────
langchain>=1.0.0
langchain-google-genai>=4.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-text-splitters>=1.0.0

# ── Google Generative AI SDK ──────────────────────────────────
google-generativeai>=0.8.0

# ── Vector Database ────────────────────────────────────────────
chromadb>=1.0.0

# ── Embeddings ─────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ── Environment & Config ──────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ──────────────────────────────────────────────────
pydantic>=2.0.0

# ── Testing ────────────────────────────────────────────────────
pytest>=8.0.0
